Madpack: Move unit tests + refactor minor code
1. Unit tests for get_rev_num and is_rev_gte moved to the correct
location
2. GPDB 5/6 version extraction made more general
3. Bare except in get_rev_num replaced with specific exceptions
